Resuscitated plate2tiles, which had bitrotted during the index update.  Also squashed some minor compiler warnings and added Int16 as a valid pixel type in preparation for ingesting MOLA and HRSC terrain.

+++ src/vw/Plate/ToastPlateManager.cc
@@ -160,7 +160,7 @@ namespace platefile {
                                                                                                        int level, 
                                                                                                        int transaction_id) const;
   template 
-  ImageView<PixelGrayA<uint8> > vw::platefile::ToastPlateManager<PixelGrayA<int16> >::fetch_child_tile(int col, int row, 
+  ImageView<PixelGrayA<int16> > vw::platefile::ToastPlateManager<PixelGrayA<int16> >::fetch_child_tile(int col, int row, 
                                                                                                        int level, 
                                                                                                        int transaction_id) const;
   template
